+++ src/scripts/train/OLMo-1B.py
@@ -32,6 +32,7 @@ from olmo_core.train.callbacks import (
     ConfigSaverCallback,
     GPUMemoryMonitorCallback,
     GradClipperCallback,
+    ProfilerCallback,
     SchedulerCallback,
     SpeedMonitorCallback,
     WandBCallback,
@@ -176,6 +177,7 @@ def build_config(run_name: str, cluster: str, overrides: List[str]) -> Experimen
             ),
         )
         .with_callback("config_saver", ConfigSaverCallback())
+        .with_callback("profiler", ProfilerCallback(enabled=False))
     )
 
     return ExperimentConfig(